+++ CHANGELOG.md
@@ -16,6 +16,12 @@ The format is based on [Keep a Changelog].
 
 ## [UNRELEASED]
 
+### Changed
+
+- The `backend.properties()` function now accepts an optional `datetime` 
+  parameter. If specified, the function returns the backend properties closest 
+  to, but older than, the specified datetime filter (\#277).
+
 ## [0.3.1] - 2019-07-23
 
 ### Changed